model.py ppgen.py
all generated pseudos get the specification in the filename. With this the dojos can be run from the base folder and the .yml files in the working folder are used and need not be copied. Moreover when all dojo's are done a table can be made from which we can read which pseudo give which result.